изменить хендлеры создания и удаления объявлений с учётом user_id в куках

+++ fastapi/src/api/endpoints/ads.py
@@ -1,7 +1,7 @@
-from fastapi import APIRouter, Depends, HTTPException
+from fastapi import APIRouter, Depends, HTTPException, Request
 from sqlalchemy import select
 from sqlalchemy_db.db import get_async_session
-from sqlalchemy_db.models import Ad
+from sqlalchemy_db.models import Ad, User
 from sqlalchemy.ext.asyncio import AsyncSession
 from src.schemas import AdOut, AdDetailOut, AdCreate, BaseResponse
 
@@ -17,7 +17,7 @@ async def get_list_ads(session: AsyncSession = Depends(get_async_session)) -> li
 async def get_ad_detail(
     ad_id: str,
     session: AsyncSession = Depends(get_async_session)
-    ) -> AdOut:
+) -> AdOut:
     ad = await session.get(Ad, ad_id)
     if not ad:
         raise HTTPException(404, detail="Объявление не найдено")
@@ -28,7 +28,7 @@ async def get_ad_detail(
 async def get_ad_detail(
     ad_id: str,
     session: AsyncSession = Depends(get_async_session)
-    ) -> AdDetailOut:
+) -> AdDetailOut:
     ad = await session.get(Ad, ad_id)
 
     if not ad:
@@ -40,10 +40,19 @@ async def get_ad_detail(
 
 @router.post("/", summary="Разместить объявление")
 async def create_ad(
+    request: Request,
     data: AdCreate,
     session: AsyncSession = Depends(get_async_session)
-    ) -> str:
-    ad = Ad(**data.model_dump())
+) -> str:
+    owner_id = request.cookies.get('user_id')
+    if not owner_id:
+        raise HTTPException(status_code=403, detail="Вы не авторизованы, войдите в систему")
+    
+    user = await session.get(User, owner_id)
+    if not user:
+        raise HTTPException(status_code=403, detail="Пользователь не найден")
+
+    ad = Ad(**data.model_dump(), owner_id=owner_id)
     session.add(ad)
     await session.commit()
     await session.refresh(ad)
@@ -51,14 +60,22 @@ async def create_ad(
 
 @router.delete("/{ad_id}", summary="Удаление объявления")
 async def delete_ad(
+    request: Request,
     ad_id: str,
     session: AsyncSession = Depends(get_async_session)
 ) -> BaseResponse:
+    owner_id = request.cookies.get('user_id')
+    if not owner_id:
+        raise HTTPException(status_code=403, detail="Вы не авторизованы, войдите в систему")
+    
     ad = await session.get(Ad, ad_id)
 
     if not ad:
         raise HTTPException(status_code=404, detail="Объявление не найдено")
-    
+
+    if ad.owner_id != owner_id:
+        raise HTTPException(status_code=403, detail="Вы не можете удалить это объявление")
+
     await session.delete(ad)
     await session.commit()
     return {"status_code": 200, "detail": "Объявление успешно удалено"}
